Car-Agent/requirements.txt
google-genai
python-dotenv
fastapi
uvicorn
pydantic
requests
